refactor: new caller from interfaces
